♻️ Move fileTypeDetail and versionState to enums
