+++ src/common/fileUtils.js
@@ -1,5 +1,4 @@
 import {KF_STUDY_API} from './globals';
-
 import * as stringSimilarity from 'string-similarity';
 
 // Compare date of file versions based on their createdAt time. (Latest first)
@@ -41,39 +40,6 @@ export const downloadFile = (
     const url = `${KF_STUDY_API}${resp.data.signedUrl.url}`;
     window.location.href = url;
   });
-};
-
-// Store file type title, description and icon
-export const fileTypeDetail = {
-  SHM: {
-    icon: 'shipping',
-    title: 'Shipping Manifest',
-    description: 'File type description goes here...',
-  },
-  CLN: {
-    icon: 'hospital',
-    title: 'Clinical/Phenotype Data',
-    description: 'File type description goes here...',
-  },
-  SEQ: {
-    icon: 'dna',
-    title: 'Sequencing Manifest',
-    description: 'File type description goes here...',
-  },
-  OTH: {
-    icon: 'question',
-    title: 'Other',
-    description: 'File type description goes here...',
-  },
-};
-
-// Store version state title and color
-export const versionState = {
-  PEN: {title: 'Pending review', labelColor: 'orange'},
-  APP: {title: 'Approved', labelColor: 'teal'},
-  CHN: {title: 'Changes needed', labelColor: 'red'},
-  PRC: {title: 'Processed', labelColor: 'blue'},
-  UPD: {title: 'Updated', labelColor: 'blue'},
 };
 
 // Sort file versions based on the version createdAt date (Latest first)
