+++ src/components/ListFilterBar/ListFilterBar.js
@@ -4,12 +4,11 @@ import {Icon, Button, Dropdown, Input, Segment} from 'semantic-ui-react';
 import Badge from '../Badge/Badge';
 import {
   fileLatestStatus,
-  versionState,
-  fileTypeDetail,
   createDateSort,
   modifiedDateSort,
   defaultSort,
 } from '../../common/fileUtils';
+import {versionState, fileTypeDetail} from '../../common/enums';
 
 /**
  * Filter Bar for Study Files, returns filtered list in "filteredList" render prop
