+++ src/modals/VersionInfoModal.js
@@ -2,12 +2,8 @@ import React from 'react';
 import Badge from '../components/Badge/Badge';
 import AvatarTimeAgo from '../components/AvatarTimeAgo/AvatarTimeAgo';
 import {Button, Header, Modal, Icon, Grid, Label} from 'semantic-ui-react';
-import {
-  downloadFile,
-  fileTypeDetail,
-  formatFileSize,
-  lengthLimit,
-} from '../common/fileUtils';
+import {downloadFile, formatFileSize, lengthLimit} from '../common/fileUtils';
+import {fileTypeDetail} from '../common/enums';
 
 const VersionInfoModal = ({
   studyId,
